logml {bridgesampling}	R Documentation

Log Marginal Likelihoods from Bridge Objects

Description

Generic function that returns log marginal likelihood from bridge objects. For objects of class "bridge_list", which contains multiple log marginal likelihoods, fun is performed on the vector and its result returned.

Usage


logml(x, ...)

## S3 method for class 'bridge'
logml(x, ...)

## S3 method for class 'bridge_list'
logml(x, fun = median, ...)

Arguments

x
Object of class "bridge" or "bridge_list" as returned from bridge_sampler.

...
Further arguments passed to fun.

fun
Function which returns a scalar value and is applied to the logml vector of "bridge_list" objects. Default is median.

Value

scalar numeric

[Package bridgesampling version 0.8-1 Index]